requirements.txt
# Core
fastapi
uvicorn[standard]
python-dotenv
pydantic
httpx

# LangChain
langchain
langchain-openai
langchain-community
langchain-core
langchain-chroma
langchain-text-splitters
langchain_pymupdf4llm

# Vector Store
chromadb

# Testing
pytest
ragas

# Development
ipykernel
